MIRPRapp/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.8.1
astroid==2.4.2
async-timeout==4.0.2
attrs==22.1.0
banal==1.0.6
certifi==2022.5.18.1
charset-normalizer==2.1.1
coverage==5.3
cycler==0.11.0
dataloader==2.0
dataset==1.5.2
datasets==2.6.1
dill==0.3.5.1
filelock==3.8.0
fonttools==4.32.0
frozenlist==1.3.1
fsspec==2022.10.0
greenlet==1.1.3.post0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
isort==5.5.3
kiwisolver==1.4.2
lazy-object-proxy==1.4.3
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib==3.5.1
mccabe==0.6.1
multidict==6.0.2
multiprocess==0.70.14
numpy==1.23.3
packaging==21.3
pandas==1.5.0
Pillow==9.1.0
pyarrow==10.0.0
pygame==2.0.0
pylint==2.6.0
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.5
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
responses==0.18.0
scipy==1.8.0
six==1.15.0
SQLAlchemy==1.4.42
texttable==1.6.3
toml==0.10.1
torch==1.13.0
tqdm==4.64.1
transforms==0.1
typing-extensions==4.4.0
urllib3==1.26.12
wrapt==1.12.1
xlrd==2.0.1
xxhash==3.1.0
yarl==1.8.1
zipp==3.10.0

torchvision~=0.14.0